feat(http sources): allow auth enrichment in all http server sources
